# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/calibration/queue": {
      "get": {
        "operationId": "CalibrationAdminController_queue",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "unreviewed",
              "type": "string",
              "enum": ["unreviewed", "mine", "all"]
            }
          },
          {
            "name": "flagged",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationQueueResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": ["admin-calibration"]
      }
    },
    "/api/admin/calibration/flags": {
      "get": {
        "operationId": "CalibrationAdminController_flags",
        "parameters": [
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "schema": {
              "default": "unreviewed",
              "type": "string",
              "enum": ["unreviewed", "mine", "all"]
            }
          },
          {
            "name": "flagged",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 10,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationFlagsResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": ["admin-calibration"]
      }
    },
...
    "/api/admin/calibration/answers/{id}": {
      "get": {
        "operationId": "CalibrationAdminController_answer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationAnswerDto_Output"
                }
              }
            }
          },
          "403": {
            "description": "Not a reviewer"
          },
          "404": {
            "description": "No such answer, or not one this reviewer may see"
          }
        },
        "tags": ["admin-calibration"]
      }
    },
    "/api/admin/calibration/answers/{id}/score": {
      "post": {
        "operationId": "CalibrationAdminController_score",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationScoreInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationScoreSavedDto_Output"
                }
              }
            }
          },
          "404": {
            "description": "No such answer, or not one this reviewer may see"
          }
        },
        "tags": ["admin-calibration"]
      }
    },
    "/api/interviews": {
      "post": {
        "operationId": "InterviewsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "No role or level to use (`profile_required`), or the role does not offer that level, variant or question type (`level_not_offered`, `stack_not_offered`)"
          },
          "404": {
            "description": "No such published role, level or variant"
          },
          "409": {
            "description": "Nothing published to ask (code `no_questions_available`)"
          },
          "429": {
            "description": "Too many interviews started (code `rate_limited`)"
          }
        },
        "tags": ["interviews"]
      },
      "get": {
        "operationId": "InterviewsController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewListResponseDto_Output"
                }
              }
            }
          }
        },
        "tags": ["interviews"]
      }
    },
...
  },
  "info": {
    "title": "Readi API",
    "description": "",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "QuestionType_Output": {
        "type": "string",
        "enum": ["behavioral", "technical", "scenario", "test_design"]
      },
...
      "QuestionType": {
        "type": "string",
        "enum": ["behavioral", "technical", "scenario", "test_design"]
      },
...
      "CalibrationQueueItem_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "question_slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "rubric_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "nullable": true
          },
          "level": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "nullable": true
          },
          "criterion_count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "review_count": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "reviewed_by_me": {
            "type": "boolean"
          },
          "flagged": {
            "type": "boolean"
          },
          "answered_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "question_slug",
          "rubric_name",
          "role",
          "level",
          "criterion_count",
          "review_count",
          "reviewed_by_me",
          "flagged",
          "answered_at"
        ],
        "additionalProperties": false
      },
      "CalibrationQueueResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalibrationQueueItem_Output"
            }
          },
          "next_cursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "nullable": true
          },
          "empty_because": {
            "type": "string",
            "enum": ["nothing_to_review", "no_consent", "staff_answers_only", null],
            "default": null,
            "nullable": true
          }
        },
        "required": ["items", "next_cursor", "empty_because"],
        "additionalProperties": false
      },
      "CalibrationFlagItem_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "question_slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "flags": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 60
            }
          },
          "answered_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": ["id", "question_slug", "flags", "answered_at"],
        "additionalProperties": false
      },
      "CalibrationFlagsResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CalibrationFlagItem_Output"
            }
          },
          "next_cursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "nullable": true
          },
          "phrases": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "phrase": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 60
                },
                "answers": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 9007199254740991
                }
              },
              "required": ["phrase", "answers"],
              "additionalProperties": false
            }
          }
        },
        "required": ["items", "next_cursor", "phrases"],
        "additionalProperties": false
      },
...
      "CalibrationCriterion_Output": {
        "type": "object",
        "properties": {
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "dimension": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "maxLength": 300
          },
          "weight": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100
          },
          "levels": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "minLength": 1,
              "maxLength": 300
            }
          }
        },
        "required": ["position", "dimension", "description", "weight", "levels"],
        "additionalProperties": false
      },
      "CalibrationTurn_Output": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "speaker": {
            "type": "string",
            "enum": ["interviewer", "candidate"]
          },
          "follow_up_index": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          }
        },
        "required": ["seq", "speaker", "follow_up_index", "text"],
        "additionalProperties": false
      },
      "CalibrationCriterionScore_Output": {
        "type": "object",
        "properties": {
          "criterion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "score": {
            "type": "integer",
            "minimum": 0,
            "maximum": 4
          },
          "max_score": {
            "type": "number",
            "enum": [4]
          },
          "evidence": {
            "maxItems": 3,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 400
            }
          },
          "reasoning": {
            "type": "string",
            "minLength": 1,
            "maxLength": 400,
            "nullable": true
          }
        },
        "required": ["criterion", "score", "max_score", "evidence", "reasoning"],
        "additionalProperties": false
      },
      "CalibrationAnswerDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "question_slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "question_prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "question_context": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000,
            "nullable": true
          },
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "nullable": true
          },
          "level": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140,
            "nullable": true
          },
          "rubric_slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "rubric_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140
          },
          "criteria": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CalibrationCriterion_Output"
            }
          },
          "exchange": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/CalibrationTurn_Output"
            }
          },
          "evidence_flags": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 60
            }
          },
          "my_score": {
            "type": "object",
            "properties": {
              "criteria": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CalibrationCriterionScore_Output"
                }
              },
              "note": {
                "type": "string",
                "minLength": 1,
                "maxLength": 2000,
                "nullable": true
              },
              "created_at": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
              }
            },
            "required": ["criteria", "note", "created_at"],
            "additionalProperties": false,
            "nullable": true
          }
        },
        "required": [
          "id",
          "question_slug",
          "question_prompt",
          "question_context",
          "role",
          "level",
          "rubric_slug",
          "rubric_name",
          "criteria",
          "exchange",
          "evidence_flags",
          "my_score"
        ],
        "additionalProperties": false
      },
      "CalibrationScoreInputDto": {
        "type": "object",
        "properties": {
          "criteria": {
            "minItems": 2,
            "maxItems": 8,
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "criterion": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "score": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 4
                },
                "evidence": {
                  "default": [],
                  "maxItems": 3,
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 400
                  }
                },
                "reasoning": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 400,
                  "default": null,
                  "nullable": true
                }
              },
              "required": ["criterion", "score"]
            }
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "default": null,
            "nullable": true
          }
        },
        "required": ["criteria"]
      },
      "CalibrationScoreSavedDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": ["id"],
        "additionalProperties": false
      },
      "CreateInterviewRequestDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "level": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "stack": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "nullable": true
          },
          "types": {
            "minItems": 1,
            "maxItems": 4,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionType"
            }
          },
          "minutes": {
            "type": "number",
            "enum": [15, 30]
          },
          "is_diagnostic": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": ["minutes"]
      },
      "InterviewState_Output": {
        "type": "string",
        "enum": ["intro", "question", "follow_up", "candidate_questions", "wrap_up", "ended"]
      },
      "InterviewStatus_Output": {
        "type": "string",
        "enum": ["in_progress", "completed", "abandoned"]
      },
      "InterviewMode_Output": {
        "type": "string",
        "enum": ["text", "voice"]
      },
      "InterviewPersona_Output": {
        "type": "string",
        "enum": ["friendly"]
      },
      "InterviewCatalogueRef_Output": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 140
          }
        },
        "required": ["slug", "name"],
        "additionalProperties": false
      },
      "CandidateSessionQuestion_Output": {
        "type": "object",
        "properties": {
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "type": {
            "$ref": "#/components/schemas/QuestionType_Output"
          },
          "topic": {
            "type": "object",
            "properties": {
              "slug": {
                "type": "string",
                "minLength": 1,
                "maxLength": 80,
                "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
              },
              "name": {
                "type": "string",
                "minLength": 1,
                "maxLength": 140
              },
              "description": {
                "type": "string",
                "minLength": 1,
                "maxLength": 400,
                "nullable": true
              },
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              }
            },
            "required": ["slug", "name", "description", "id"],
            "additionalProperties": false
          },
          "prompt": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "context": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000,
            "nullable": true
          },
          "asked_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": ["position", "type", "topic", "prompt", "context", "asked_at"],
        "additionalProperties": false
      },
      "CandidateTurn_Output": {
        "type": "object",
        "properties": {
          "seq": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          },
          "speaker": {
            "$ref": "#/components/schemas/TurnSpeaker_Output"
          },
          "state": {
            "$ref": "#/components/schemas/InterviewState_Output"
          },
          "question_position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991,
            "nullable": true
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          }
        },
        "required": ["seq", "speaker", "state", "question_position", "text", "at"],
        "additionalProperties": false
      },
      "TurnSpeaker_Output": {
        "type": "string",
        "enum": ["interviewer", "candidate"]
      },
      "InterviewSessionResponseDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "state": {
            "$ref": "#/components/schemas/InterviewState_Output"
          },
          "status": {
            "$ref": "#/components/schemas/InterviewStatus_Output"
          },
          "mode": {
            "$ref": "#/components/schemas/InterviewMode_Output"
          },
          "persona": {
            "$ref": "#/components/schemas/InterviewPersona_Output"
          },
          "is_diagnostic": {
            "type": "boolean"
          },
          "planned_minutes": {
            "type": "number",
            "enum": [15, 30]
          },
          "role": {
            "$ref": "#/components/schemas/InterviewCatalogueRef_Output"
          },
          "level": {
            "$ref": "#/components/schemas/InterviewCatalogueRef_Output"
          },
          "stack": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewCatalogueRef_Output"
              }
            ],
            "nullable": true
          },
          "types": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionType_Output"
            }
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "ended_at": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "nullable": true
          },
          "question_budget": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "max_follow_ups": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateSessionQuestion_Output"
            }
          },
          "turns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateTurn_Output"
            }
          }
        },
        "required": [
          "id",
          "state",
          "status",
          "mode",
          "persona",
          "is_diagnostic",
          "planned_minutes",
          "role",
          "level",
          "stack",
          "types",
          "started_at",
          "ends_at",
          "ended_at",
          "question_budget",
          "max_follow_ups",
          "questions",
          "turns"
        ],
        "additionalProperties": false
      },
...
      "InterviewListResponseDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewSummary_Output"
            }
          },
          "next_cursor": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "nullable": true
          }
        },
        "required": ["items", "next_cursor"],
        "additionalProperties": false
      },
...
    }
  }
}